substitute some boost function with std equivalent

+++ src/HdrWizard/AutoAntighosting.cpp
@@ -134,10 +134,10 @@ int findIndex(const float *data, int size) {
     return idx;
 }
 
-typedef vector<float> BufferF;
+typedef vector<double> BufferF;
 
 inline float hueMean(const BufferF &data) {
-    return std::accumulate(data.begin(), data.end(), 0.0f) / data.size();
+    return std::accumulate(data.begin(), data.end(), 0.0d) / data.size();
 }
 
 void hueSquaredMean(const HdrCreationItemContainer &data, vector<float> &HE) {
